move DetectHomeDir to fileutil
Both cliutil and fileutil contain similar code to detect the home
directory; this combines the two functions into a single function.
This will ensure we consistently use the same home dir detection.
